requirements.txt
slackclient==2.9.0
PyVimeo==1.1.0
